+++ nanoCAT/asa/md_asa.py
@@ -19,15 +19,14 @@ API
 """
 
 from os.path import join
-from typing import Iterable, Tuple, Any, Type, Generator, Iterator, TypeVar
-from itertools import chain, repeat
+from typing import Iterable, Tuple, Any, Type, Generator
+from itertools import chain
 
 import numpy as np
 
-from rdkit import Chem
 from scm.plams import Settings, Molecule, Cp2kJob, Units
 from scm.plams.core.basejob import Job
-import scm.plams.interfaces.molecule.rdkit as molkit
+from scm.plams.interfaces.molecule.rdkit import add_Hs
 
 from FOX import (
     get_non_bonded, get_intra_non_bonded, get_bonded, MultiMolecule, PSFContainer, PRMContainer
@@ -35,14 +34,16 @@ from FOX import (
 
 from CAT.jobs import job_md
 from CAT.mol_utils import round_coords
-from CAT.attachment.qd_opt_ff import qd_opt_ff, _constrain_charge
+from CAT.attachment.qd_opt_ff import qd_opt_ff
 
 from .asa_frag import get_asa_fragments
 from ..ff.ff_assignment import run_match_job
 
 
 def get_asa_md(mol_list: Iterable[Molecule], jobs: Tuple[Type[Job], ...],
-               settings: Tuple[Settings, ...], **kwargs: Any) -> np.ndarray:
+               settings: Tuple[Settings, ...], iter_start: int = 500,
+               scale_elstat: float = 0.0, scale_lj: float = 1.0,
+               **kwargs: Any) -> np.ndarray:
     r"""Perform an activation strain analyses (ASA) along an molecular dynamics (MD) trajectory.
 
     The ASA calculates the (ensemble-averaged) interaction, strain and total energy.
@@ -57,6 +58,18 @@ def get_asa_md(mol_list: Iterable[Molecule], jobs: Tuple[Type[Job], ...],
 
     settings : :class:`tuple` [|plams.Settings|]
         A tuple containing a single |plams.Settings| instance.
+
+    iter_start : :class:`int`
+        The MD iteration at which the ASA will be started.
+        All preceding iteration are disgarded, treated as pre-equilibration steps.
+
+    scale_elstat : :class:`float`
+        Scaling factor to apply to all 1,4-nonbonded electrostatic interactions.
+        Serves the same purpose as the cp2k ``EI_SCALE14`` keyword.
+
+    scale_lj : :class:`float`
+        Scaling factor to apply to all 1,4-nonbonded Lennard-Jones interactions.
+        Serves the same purpose as the cp2k ``VDW_SCALE14`` keyword.
 
     \**kwargs : :data:`Any<typing.Any>`
         Further keyword arguments for ensuring signature compatiblity.
@@ -87,26 +100,29 @@ def get_asa_md(mol_list: Iterable[Molecule], jobs: Tuple[Type[Job], ...],
         count = mol_len * 5
 
     # Extract all energies and ligand counts
-    iterator = chain.from_iterable(md_generator(mol_list, job, s))
+    iterator = chain.from_iterable(md_generator(mol_list, job, s,
+                                                iter_start=iter_start,
+                                                scale_elstat=scale_elstat,
+                                                scale_lj=scale_lj))
+
     E = np.fromiter(iterator, count=count, dtype=float)
     E.shape = shape
-    E[:, :4] *= Units.conversion_ratio('au', 'kcal/mol')
+    E[:, :4] *= Units.conversion_ratio('hartree', 'kcal/mol')
 
     # Calculate (and return) the interaction, strain and total energy
     E_int = E[:, 0]
-    E_strain = np.sum(E[:, 1:3], axis=1) - np.product(E[:, 3:], axis=1)
+    E_strain = E[:, 1:3].sum(axis=1) - E[:, 3:].prod(axis=1)
 
-    ret = np.array([E_int, E_strain, E_int + E_strain]).T
-    return ret
+    return np.array([E_int, E_strain, E_int + E_strain]).T
 
 
 MATCH_SETTINGS = Settings({'input': {'forcefield': 'top_all36_cgenff'}})
-KCAL2AU: float = Units.conversion_ratio('kcal/mol', 'hartree')  # kcal/mol to hartree
 Tuple5 = Tuple[float, float, float, float, int]
 
 
 def md_generator(mol_list: Iterable[Molecule], job: Type[Job],
-                 settings: Settings) -> Generator[Tuple5, None, None]:
+                 settings: Settings, iter_start: int = 500,
+                 scale_elstat: float = 0.0, scale_lj: float = 1.0) -> Generator[Tuple5, None, None]:
     """Iterate over an iterable of molecules; perform an MD followed by an ASA.
 
     The various energies are averaged over all molecules in the MD-trajectory.
@@ -123,6 +139,18 @@ def md_generator(mol_list: Iterable[Molecule], job: Type[Job],
     settings : :class:`tuple` [|plams.Settings|]
         CP2K job settings for **job**.
 
+    iter_start : :class:`int`
+        The MD iteration at which the ASA will be started.
+        All preceding iteration are disgarded, treated as pre-equilibration steps.
+
+    scale_elstat : :class:`float`
+        Scaling factor to apply to all 1,4-nonbonded electrostatic interactions.
+        Serves the same purpose as the cp2k ``EI_SCALE14`` keyword.
+
+    scale_lj : :class:`float`
+        Scaling factor to apply to all 1,4-nonbonded Lennard-Jones interactions.
+        Serves the same purpose as the cp2k ``VDW_SCALE14`` keyword.
+
     Returns
     -------
     4x :class:`float` and 1x :class:`int`
@@ -131,61 +159,85 @@ def md_generator(mol_list: Iterable[Molecule], job: Type[Job],
         * The inter-ligand non-bonded interaction
         * The intra-ligand non-bonded interaction
         * The intra-ligand bonded interaction
-        * The energy of a single optimized ligand
+        * The energy of a single optimized ligand (bonded & non-bonded interactions)
         * The number of ligands
 
     """
     for mol in mol_list:
+        # Identify the fragments
+        ligands, _ = get_asa_fragments(mol)
+        lig = ligands[0]
+        lig.round_coords()
+        lig_count = len(ligands)
+
         # Run the MD job
-        results = qd_opt_ff(mol, job, settings, name='QD_MD', job_func=Molecule.job_md)
-        if results.job.status == 'crashed':
+        md_results = qd_opt_ff(mol, job, settings, name='QD_MD', job_func=Molecule.job_md)
+        if md_results.job.status == 'crashed':
             yield np.nan, np.nan, np.nan, np.nan, 0
+            continue
 
-        multi_mol = MultiMolecule.from_xyz(results['cp2k-pos-1.xyz'])
-        psf = PSFContainer.read(results['QD_MD.psf'])
+        md_trajec = MultiMolecule.from_xyz(md_results['cp2k-pos-1.xyz'])[iter_start:]
+        psf_charged = PSFContainer.read(md_results['QD_MD.psf'])
 
-        # Switch to the neutral parameters
-        frags, _ = get_asa_fragments(mol)
-        frag = frags[0]
-        frag.round_coords()
-        frag_count = len(frags)
-        frag_neutral = frag.copy()
-        frag_neutral[len(mol) - mol.properties.indices[-1] - 1].properties.charge = 0
-        frag_neutral = add_Hs(frag_neutral, forcefield='uff')
-        run_match_job(frag_neutral, MATCH_SETTINGS)
+        # Optimize a single ligand
+        opt_results = qd_opt_ff(lig, job, _md2opt(settings), new_psf=True, name='ligand_opt')
+        if opt_results.job.status == 'crashed':
+            yield np.nan, np.nan, np.nan, np.nan, 0
+            continue
 
-        # Update the PSFContainer
-        psf_neutral = psf.copy()
-        symbol_list = [at.properties.symbol for at in frag_neutral.atoms[:-1]] * frag_count
-        charge_list = [at.properties.charge_float for at in frag_neutral.atoms[:-1]] * frag_count
-        psf_neutral.atom_type.loc[psf_neutral.residue_name == 'LIG'] = symbol_list
-        psf_neutral.charge.loc[psf_neutral.residue_name == 'LIG'] = charge_list
+        # Prepare arguments for the intra-ligand interactions
+        lig_opt = MultiMolecule.from_Molecule(lig)
+        prm_charged = PRMContainer.read(opt_results['ligand_opt.prm'])
+        psf_lig = join(opt_results.job.path, 'ligand_opt.psf')
 
-        psf_neutral.charge.loc[mol.properties.indices] += frag_neutral[-1].properties.charge_float
-        psf_neutral.charge.loc[psf_neutral.residue_name == 'COR'] = 0.0
+        # Prepare arguments for the inter-ligand interactions
+        lig_neutral = _get_neutral_frag(lig)
+        prm_neutral = PRMContainer.read(lig_neutral.properties.prm)
+        psf_neutral = _get_neutral_psf(psf_charged, lig_neutral,
+                                       lig_count, mol.properties.indices)
 
-        # Calculate all inter-ligand interactions
-        prm_neutral = PRMContainer.read(frag_neutral.properties.prm)
-        inter_nb = _inter_nonbonded(multi_mol, None, psf_neutral, prm_neutral)
+        # Inter-ligand interaction
+        inter_nb = _inter_nonbonded(md_trajec, None, psf_neutral, prm_neutral)
 
-        # Calculate all intra-ligand interactions
-        prm_charge = PRMContainer.read(results['QD_MD.prm'])
-        intra_nb = _intra_nonbonded(multi_mol, psf, prm_charge)
-        inter_bond = _inter_bonded(multi_mol, psf, prm_charge)
+        # Intra-ligand interaction
+        intra_nb = _intra_nonbonded(md_trajec, psf_charged, prm_charged)
+        intra_bond = _inter_bonded(md_trajec, psf_charged, prm_charged)
 
-        # Optimize an (individual) ligand
-        results = qd_opt_ff(frag, job, md2opt(settings), new_psf=True, name='Ligand_opt')
+        # Intra-ligand interaction within a single optimized ligand
+        frag_opt = _intra_nonbonded(lig_opt, psf_lig, prm_charged)
+        frag_opt += _inter_bonded(lig_opt, psf_lig, prm_charged)
 
-        # Calculate the optimized ligand energy
-        frag_multi = MultiMolecule.from_Molecule(frag)
-        psf_lig = join(results.job.path, 'Ligand_opt.psf')
-        frag_opt = _intra_nonbonded(frag_multi, psf_lig, prm_charge)
-        frag_opt += _inter_bonded(frag_multi, psf_lig, prm_charge)
-
-        yield inter_nb, intra_nb, inter_bond, frag_opt, frag_count
+        yield inter_nb, intra_nb, intra_bond, frag_opt, lig_count
 
 
-def md2opt(s: Settings) -> Settings:
+def _get_neutral_frag(frag: Molecule) -> Molecule:
+    """Return a neutral fragment for :func:`md_generator`."""
+    frag_neutral = frag.copy()
+    for at in frag_neutral:
+        if at.properties.anchor:
+            at.properties.charge = 0
+            break
+    frag_neutral = add_Hs(frag_neutral, forcefield='uff')
+    run_match_job(frag_neutral, MATCH_SETTINGS)
+    return frag_neutral
+
+
+def _get_neutral_psf(psf: PSFContainer, frag_neutral: Molecule, frag_count: int,
+                     anchor_idx: np.ndarray) -> PSFContainer:
+    """Return a net-neutral :class:`PSFContainer` for :func:`md_generator`."""
+    psf_neutral = psf.copy()
+    symbol_list = [at.properties.symbol for at in frag_neutral.atoms[:-1]] * frag_count
+    charge_list = [at.properties.charge_float for at in frag_neutral.atoms[:-1]] * frag_count
+
+    psf_neutral.atom_type.loc[psf_neutral.residue_name == 'LIG'] = symbol_list
+    psf_neutral.charge.loc[psf_neutral.residue_name == 'LIG'] = charge_list
+
+    psf_neutral.charge.loc[anchor_idx] += frag_neutral[-1].properties.charge_float
+    psf_neutral.charge.loc[psf_neutral.residue_name == 'COR'] = 0.0
+    return psf_neutral
+
+
+def _md2opt(s: Settings) -> Settings:
     """Convert CP2K MD settings to CP2K geometry optimization settings."""
     s2 = s.copy()
     del s2.input.motion.md
@@ -197,8 +249,8 @@ def md2opt(s: Settings) -> Settings:
     return s2
 
 
-def _inter_nonbonded(multi_mol: MultiMolecule, s: Settings,
-                     psf: PSFContainer, prm: PRMContainer) -> float:
+def _inter_nonbonded(multi_mol: MultiMolecule, s: Settings, psf: PSFContainer,
+                     prm: PRMContainer) -> float:
     """Collect all inter-ligand non-bonded interactions."""
     # Manually calculate all inter-ligand, ligand/core & core/core interactions
     df = get_non_bonded(multi_mol, psf=psf, prm=prm, cp2k_settings=s)
@@ -212,53 +264,16 @@ def _inter_nonbonded(multi_mol: MultiMolecule, s: Settings,
     return df.values.sum()
 
 
-def _intra_nonbonded(multi_mol: MultiMolecule, psf: PSFContainer, prm: PRMContainer) -> float:
+def _intra_nonbonded(multi_mol: MultiMolecule, psf: PSFContainer, prm: PRMContainer,
+                     scale_elstat: float = 0.0, scale_lj: float = 1.0) -> float:
     """Collect all intra-ligand non-bonded interactions."""
-    return get_intra_non_bonded(multi_mol, psf=psf, prm=prm).values.sum()
+    df = get_intra_non_bonded(multi_mol, psf, prm,
+                              scale_elstat=scale_elstat,
+                              scale_lj=scale_lj)
+    return df.values.sum()
 
 
 def _inter_bonded(multi_mol: MultiMolecule, psf: PSFContainer, prm: PRMContainer) -> float:
     """Collect all intra-ligand bonded interactions."""
     E_tup = get_bonded(multi_mol, psf, prm)  # bonds, angles, dihedrals, impropers
     return sum((series.sum() if series is not None else 0.0) for series in E_tup)
-
-
-# Temporary replacement for molkit's add_Hs() function until the function is fixed
-# https://github.com/SCM-NV/PLAMS/pull/77
-
-def add_Hs(mol, forcefield=None, return_rdmol=False):
-    """Add hydrogens to protein molecules read from PDB.
-
-    Makes sure that the hydrogens get the correct PDBResidue info.
-
-    :param mol: Molecule to be protonated
-    :type mol: |Molecule| or rdkit.Chem.Mol
-    :param str forcefield: Specify 'uff' or 'mmff' to apply forcefield based
-        geometry optimization on new atoms.
-    :param bool return_rdmol: return a RDKit molecule if true, otherwise a PLAMS molecule
-    :return: A molecule with explicit hydrogens added
-    :rtype: |Molecule| or rdkit.Chem.Mol
-    """
-    mol = molkit.to_rdmol(mol)
-    retmol = Chem.AddHs(mol)
-    for atom in retmol.GetAtoms():
-        if atom.GetPDBResidueInfo() is None and atom.GetSymbol() == 'H':
-            bond = atom.GetBonds()[0]
-            if bond.GetBeginAtom().GetIdx() == atom.GetIdx:
-                connected_atom = bond.GetEndAtom()
-            else:
-                connected_atom = bond.GetBeginAtom()
-            try:
-                ResInfo = connected_atom.GetPDBResidueInfo()
-                if ResInfo is None:
-                    continue  # Segmentation faults are raised if ResInfo is None
-                atom.SetMonomerInfo(ResInfo)
-                atomname = 'H' + atom.GetPDBResidueInfo().GetName()[1:]
-                atom.GetPDBResidueInfo().SetName(atomname)
-            except Exception:
-                pass
-
-    unchanged = molkit.gen_coords_rdmol(retmol)
-    if forcefield:
-        molkit.optimize_coordinates(retmol, forcefield, fixed=unchanged)
-    return retmol if return_rdmol else molkit.from_rdmol(retmol)